bluetooth: host: Expose structs and prototypes in public headers
In order for custom settings backends to be able to extract settings
values for efficient storage, expose structs from keys.h and gatt.c, as
well as bt_settings_encode_key and bt_settings_decode_key prototypes.
